scan.l: Support @pragma {start,end}-ignore-deprecations
Mostly to have something to talk about. It's rather verbose in usage.

Benjamin brought up that it would be better to track ranges to be
ignored and decide whether deprecations should be output likely
post parsing instead of during parsing. That would likely also better
fit with the usage analyzer's approach like "start-ignore-unsued",
though that went with the &is_used attribute instead, so not sure
it's worth it yet.